+++ src/common/api/records.api.ts
@@ -61,6 +61,19 @@ export const getMarcRecord = async ({ recordId, endpointUrl }: SingleRecord & { 
   });
 };
 
+const instanceRdfUrl = `${BIBFRAME_API_ENDPOINT}/:recordId/rdf`;
+
+export const getRdfRecord = ( recordId: string ) => {
+  const url = baseApi.generateUrl(instanceRdfUrl, {
+    name: ':recordId',
+    value: recordId,
+  });
+
+  return baseApi.request({
+    url,
+  });
+};
+
 export const getAllRecords = async ({
   pageSize = MAX_LIMIT,
   pageNumber = MAX_LIMIT,